% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab4.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{A linear regression function}
\description{
A RC class function that can handle linear regression problems like function lm()
}
\section{Fields}{

\describe{
\item{\code{beta_hat:}}{Regressions coefficients}

\item{\code{X}}{independent variables}

\item{\code{y}}{dependent variable}
}}


