% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab4.R
\docType{class}
\name{ridgereg-class}
\alias{ridgereg-class}
\alias{ridgereg}
\title{A ridge regression function}
\arguments{
\item{formula}{a formula}

\item{data}{a data frame}

\item{lambda}{a numeric}
}
\description{
A RC class function that can handle ridge regression problems similar as lm.ridge(), it also handles
the special functions print(), predict(), and coef().
}
\section{Fields}{

\describe{
\item{\code{beta_hat_ridge}}{Regressions coefficients}

\item{\code{X}}{independent variables}

\item{\code{X_norm}}{normalization X}

\item{\code{y}}{dependent variable}

\item{\code{y_hat}}{fitted values}

\item{\code{formula}}{formula to ridge regression}

\item{\code{data}}{data frame to analyse}

\item{\code{df_name}}{name of the analysed data frame}

\item{\code{lambda}}{value to tune ridge regression models}
}}


